Refactored usdm generators and extracted helper functions to usdm_utils.py; added caching for API functions
