+++ src/usdm/generate_biomedical_concepts.py
@@ -1,156 +1,12 @@
 #!/usr/bin/env python3
 from concurrent.futures import ThreadPoolExecutor
-from typing import Optional, List, Dict, Any
-import functools
-import os
-import requests
+from typing import List, Dict, Any
 
 from soa_builder.web.utils import _connect
-
-# GLobal API URL prefix
-URL_PREFIX = "https://api.library.cdisc.org/api/cosmos/v2/"
-
-
-def _build_api_headers() -> dict:
-    api_key = os.environ.get("CDISC_API_KEY") or os.environ.get(
-        "CDISC_SUBSCRIPTION_KEY"
-    )
-    subscription_key = os.environ.get("CDISC_SUBSCRIPTION_KEY") or api_key
-    headers: dict = {"Accept": "application/json"}
-    if subscription_key:
-        headers["Ocp-Apim-Subscription-Key"] = subscription_key
-    if api_key:
-        headers["Authorization"] = f"Bearer {api_key}"
-        headers["api-key"] = api_key
-    return headers
-
-
-@functools.lru_cache(maxsize=128)
-def _fetch_dss_variable_map(dss_href: str) -> Dict[str, List[str]]:
-    """Fetch a DSS href once and return {variable_name: valueList}. Cached per href."""
-    try:
-        resp = requests.get(dss_href, headers=_build_api_headers(), timeout=15)
-        if resp.status_code != 200:
-            return {}
-        data = resp.json()
-        return {v["name"]: v.get("valueList", []) for v in data.get("variables", [])}
-    except (requests.RequestException, ValueError) as e:
-        print(f"Error fetching DSS variable map: {e}")
-        return {}
-
-
-@functools.lru_cache(maxsize=256)
-def _get_dss_response_codes(
-    biomedical_concept_uid: str, variable_name: str, soa_id: int
-) -> List[str]:
-    """Return responseCodes for a single DSS variable. Cached per (bc_uid, name, soa_id)."""
-    conn = _connect()
-    cur = conn.cursor()
-    cur.execute(
-        "SELECT dss_href FROM activity_concept WHERE concept_uid=? AND soa_id=?",
-        (biomedical_concept_uid, soa_id),
-    )
-    row = cur.fetchone()
-    conn.close()
-    if not row or not row[0]:
-        return []
-    return _fetch_dss_variable_map(row[0]).get(variable_name, [])
-
-
-@functools.lru_cache(maxsize=256)
-def _get_biomedical_concept_synonyms(concept_code: str) -> List[str]:
-    """Fetch the synonyms of a biomedical concept using the CDISC API. Cached per code."""
-    url = URL_PREFIX + "mdr/bc/biomedicalconcepts/" + concept_code
-    try:
-        resp = requests.get(url, headers=_build_api_headers(), timeout=15)
-        if resp.status_code != 200:
-            return []
-        data = resp.json()
-        return data.get("synonyms", [])
-    except (requests.RequestException, ValueError) as e:
-        print(f"Error fetching biomedical concept synonyms: {e}")
-        return []
-
-
-def _get_biomedical_concept_properties(
-    soa_id: int, biomedical_concept_uid: str
-) -> Optional[Dict[str, any]]:
-    """Fetch biomedical concept properties from the database using the BiomedicalConcept_{}."""
-
-    conn = _connect()
-    cur = conn.cursor()
-    cur.execute(
-        """
-        SELECT
-            p.biomedical_concept_property_uid id,
-            p.name name,
-            p.label label,
-            p.isRequired isRequired,
-            p.datatype datatype,
-            p.biomedical_concept_uid biomedical_concept_uid,
-            a.alias_code_uid alias_code_uid,
-            a.standard_code standard_code,
-            c.code code,
-            c.code_system code_system,
-            c.code_system_version code_system_version,
-            c.decode decode
-        FROM biomedical_concept_property p
-        INNER JOIN alias_code a ON p.code = a.alias_code_uid AND p.soa_id = a.soa_id
-        INNER JOIN code c ON a.standard_code = c.code_uid AND a.soa_id = c.soa_id
-        WHERE p.soa_id = ? AND p.biomedical_concept_uid = ?
-        ORDER BY p.id;
-        """,
-        (soa_id, biomedical_concept_uid),
-    )
-    rows = cur.fetchall()
-    cur.close()
-    conn.close()
-    out: List[Dict[str, Any]] = []
-
-    for r in rows:
-        id = r[0]
-        name = r[1]
-        label = r[2]
-        isRequired = bool(r[3])
-        datatype = r[4]
-        bc_uid = r[5]
-        alias_code_uid = r[6]
-        standard_code = r[7]
-        code = r[8]
-        code_system = r[9]
-        code_system_version = r[10]
-        decode = r[11]
-
-        isEnabled = None
-        response_codes = _get_dss_response_codes(bc_uid, name, soa_id)
-
-        property = {
-            "id": id,
-            "name": name,
-            "label": label,
-            "isRequired": isRequired,
-            "isEnabled": isEnabled,
-            "datatype": datatype,
-            "responseCodes": response_codes,
-            "code": {
-                "id": alias_code_uid,
-                "extensionAttributes": [],
-                "standardCode": {
-                    "id": standard_code,
-                    "extensionAttributes": [],
-                    "code": code,
-                    "codeSystem": code_system,
-                    "codeSystemVersion": code_system_version,
-                    "decode": decode,
-                    "instanceType": "Code",
-                },
-            },
-            "notes": [],
-            "instanceType": "BiomedicalConceptProperty",
-        }
-        out.append(property)
-
-    return out
+from .usdm_utils import (
+    _get_biomedical_concept_synonyms as _get_biomedical_concept_synonyms,
+    _get_biomedical_concept_properties as _get_biomedical_concept_properties,
+)
 
 
 def build_usdm_biomedical_concepts(soa_id: int) -> List[Dict[str, Any]]:
