+++ src/usdm/generate_scheduled_activity_instances.py
@@ -1,43 +1,10 @@
 #!/usr/bin/env python3
 # Prefer absolute import; fallback to adding src/ to sys.path when run directly
 from typing import Optional, List, Dict, Any
-
-try:
-    from soa_builder.web.app import _connect  # reuse existing DB connector
-except ImportError:
-    import sys
-    from pathlib import Path
-
-    here = Path(__file__).resolve()
-    src_dir = here.parents[2] / "src"
-    if src_dir.exists() and str(src_dir) not in sys.path:
-        sys.path.insert(0, str(src_dir))
-    from soa_builder.web.app import _connect  # type: ignore
-
-
-def _nz(s: Optional[str]) -> Optional[str]:
-    s = (s or "").strip()
-    return s or None
-
-
-def _get_activity_ids(soa_id: int, encounter_uid: str) -> List[str]:
-    conn = _connect()
-    cur = conn.cursor()
-    cur.execute(
-        "SELECT a.activity_uid from activity a "
-        "INNER JOIN matrix_cells m ON a.id = m.activity_id AND a.soa_id = m.soa_id "
-        "INNER JOIN visit v ON m.visit_id = v.id AND m.soa_id = v.soa_id "
-        "INNER JOIN instances i ON v.encounter_uid = i.encounter_uid AND v.soa_id = i.soa_id "
-        "WHERE i.soa_id=? and i.encounter_uid=?",
-        (
-            soa_id,
-            encounter_uid,
-        ),
-    )
-    rows = cur.fetchall()
-    conn.close()
-    activity_uids = [r[0] for r in rows] or []
-    return activity_uids
+import sys
+from soa_builder.web.db import _connect
+from soa_builder.web.utils import _nz
+from .usdm_utils import _get_activity_ids
 
 
 def build_usdm_instances(
@@ -86,7 +53,7 @@ def build_usdm_instances(
 
     for i, r in enumerate(rows):
         (
-            id,
+            _,
             instance_uid,
             name,
             label,
@@ -132,7 +99,6 @@ if __name__ == "__main__":
     import argparse
     import json
     import logging
-    import sys
 
     logger = logging.getLogger("usdm.generate_instances")
